application/backend/requirements.txt
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.6
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
numpy==1.21.0
Pillow==10.0.0
requests==2.20.0
six==1.11.0
torch==1.11.0
torchvision==0.11.1
urllib3==1.21.1
Werkzeug==0.14.1
prometheus_flask_exporter
